+++ block/src/formats/qcow/internal/decoder.rs
