+++ block/src/formats/qcow/internal/metadata.rs
