+++ block/src/formats/qcow/internal/qcow_raw_file.rs
