+++ block/src/formats/qcow/internal/raw_file.rs
