+++ block/src/formats/qcow/internal/util.rs
