+++ block/src/formats/qcow/internal/vec_cache.rs
