+++ block/src/formats/qcow/worker/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2026 The Cloud Hypervisor Authors. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#[cfg(feature = "io_uring")]
+pub(crate) mod async_uring;
+pub(crate) mod sync;
+
+pub(crate) use super::{common, internal};
