+++ performance-metrics/src/util.rs
@@ -15,8 +15,8 @@ use std::thread;
 use std::time::Duration;
 
 use block::async_io::{AsyncIo, GuestMemoryTarget};
+use block::formats::qcow::QcowDisk;
 use block::qcow::{BackingFileConfig, ImageType, QcowFile, RawFile};
-use block::qcow_disk::QcowDisk;
 use vm_memory::{Bytes, GuestAddress, GuestMemoryMmap};
 use vmm_sys_util::eventfd::EventFd;
 use vmm_sys_util::tempfile::TempFile;
